+++ service/amplify/api.go
@@ -5087,6 +5087,27 @@ type CustomRule struct {
 	Source *string `locationName:"source" min:"1" type:"string" required:"true"`
 
 	// The status code for a URL rewrite or redirect rule.
+	//
+	// 200
+	//
+	// Represents a 200 rewrite rule.
+	//
+	// 301
+	//
+	// Represents a 301 (moved pemanently) redirect rule. This and all future requests
+	// should be directed to the target URL.
+	//
+	// 302
+	//
+	// Represents a 302 temporary redirect rule.
+	//
+	// 404
+	//
+	// Represents a 404 redirect rule.
+	//
+	// 404-200
+	//
+	// Represents a 404 rewrite rule.
 	Status *string `locationName:"status" min:"3" type:"string"`
 
 	// The target pattern for a URL rewrite or redirect rule.