+++ service/eks/api.go
@@ -2630,6 +2630,9 @@ type Cluster struct {
 	// The identity provider information for the cluster.
 	Identity *Identity `locationName:"identity" type:"structure"`
 
+	// Network configuration settings for your cluster.
+	KubernetesNetworkConfig *KubernetesNetworkConfigResponse `locationName:"kubernetesNetworkConfig" type:"structure"`
+
 	// The logging configuration for your cluster.
 	Logging *Logging `locationName:"logging" type:"structure"`
 
@@ -2718,6 +2721,12 @@ func (s *Cluster) SetIdentity(v *Identity) *Cluster {
 	return s
 }
 
+// SetKubernetesNetworkConfig sets the KubernetesNetworkConfig field's value.
+func (s *Cluster) SetKubernetesNetworkConfig(v *KubernetesNetworkConfigResponse) *Cluster {
+	s.KubernetesNetworkConfig = v
+	return s
+}
+
 // SetLogging sets the Logging field's value.
 func (s *Cluster) SetLogging(v *Logging) *Cluster {
 	s.Logging = v
@@ -2775,6 +2784,9 @@ type CreateClusterInput struct {
 
 	// The encryption configuration for the cluster.
 	EncryptionConfig []*EncryptionConfig `locationName:"encryptionConfig" type:"list"`
+
+	// The Kubernetes network configuration for the cluster.
+	KubernetesNetworkConfig *KubernetesNetworkConfigRequest `locationName:"kubernetesNetworkConfig" type:"structure"`
 
 	// Enable or disable exporting the Kubernetes control plane logs for your cluster
 	// to CloudWatch Logs. By default, cluster control plane logs aren't exported
@@ -2864,6 +2876,12 @@ func (s *CreateClusterInput) SetClientRequestToken(v string) *CreateClusterInput
 // SetEncryptionConfig sets the EncryptionConfig field's value.
 func (s *CreateClusterInput) SetEncryptionConfig(v []*EncryptionConfig) *CreateClusterInput {
 	s.EncryptionConfig = v
+	return s
+}
+
+// SetKubernetesNetworkConfig sets the KubernetesNetworkConfig field's value.
+func (s *CreateClusterInput) SetKubernetesNetworkConfig(v *KubernetesNetworkConfigRequest) *CreateClusterInput {
+	s.KubernetesNetworkConfig = v
 	return s
 }
 
@@ -3075,9 +3093,9 @@ type CreateNodegroupInput struct {
 	_ struct{} `type:"structure"`
 
 	// The AMI type for your node group. GPU instance types should use the AL2_x86_64_GPU
-	// AMI type, which uses the Amazon EKS-optimized Linux AMI with GPU support.
-	// Non-GPU instances should use the AL2_x86_64 AMI type, which uses the Amazon
-	// EKS-optimized Linux AMI. If you specify launchTemplate, and your launch template
+	// AMI type. Non-GPU instances should use the AL2_x86_64 AMI type. Arm instances
+	// should use the AL2_ARM_64 AMI type. All types use the Amazon EKS-optimized
+	// Amazon Linux 2 AMI. If you specify launchTemplate, and your launch template
 	// uses a custom AMI, then don't specify amiType, or the node group deployment
 	// will fail. For more information about using launch templates with Amazon
 	// EKS, see Launch template support (https://docs.aws.amazon.com/eks/latest/userguide/launch-templates.html)
@@ -3115,8 +3133,8 @@ type CreateNodegroupInput struct {
 	Labels map[string]*string `locationName:"labels" type:"map"`
 
 	// An object representing a node group's launch template specification. If specified,
-	// then do not specify instanceTypes, diskSize, or remoteAccess. If specified,
-	// make sure that the launch template meets the requirements in launchTemplateSpecification.
+	// then do not specify instanceTypes, diskSize, or remoteAccess and make sure
+	// that the launch template meets the requirements in launchTemplateSpecification.
 	LaunchTemplate *LaunchTemplateSpecification `locationName:"launchTemplate" type:"structure"`
 
 	// The Amazon Resource Name (ARN) of the IAM role to associate with your node
@@ -4377,6 +4395,72 @@ func (s *Issue) SetResourceIds(v []*string) *Issue {
 	return s
 }
 
+// The Kubernetes network configuration for the cluster.
+type KubernetesNetworkConfigRequest struct {
+	_ struct{} `type:"structure"`
+
+	// The CIDR block to assign Kubernetes service IP addresses from. If you don't
+	// specify a block, Kubernetes assigns addresses from either the 10.100.0.0/16
+	// or 172.20.0.0/16 CIDR blocks. We recommend that you specify a block that
+	// does not overlap with resources in other networks that are peered or connected
+	// to your VPC. The block must meet the following requirements:
+	//
+	//    * Within one of the following private IP address blocks: 10.0.0.0/8, 172.16.0.0.0/12,
+	//    or 192.168.0.0/16.
+	//
+	//    * Doesn't overlap with any CIDR block assigned to the VPC that you selected
+	//    for VPC.
+	//
+	//    * Between /24 and /12.
+	//
+	// You can only specify a custom CIDR block when you create a cluster and can't
+	// change this value once the cluster is created.
+	ServiceIpv4Cidr *string `locationName:"serviceIpv4Cidr" type:"string"`
+}
+
+// String returns the string representation
+func (s KubernetesNetworkConfigRequest) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s KubernetesNetworkConfigRequest) GoString() string {
+	return s.String()
+}
+
+// SetServiceIpv4Cidr sets the ServiceIpv4Cidr field's value.
+func (s *KubernetesNetworkConfigRequest) SetServiceIpv4Cidr(v string) *KubernetesNetworkConfigRequest {
+	s.ServiceIpv4Cidr = &v
+	return s
+}
+
+// The Kubernetes network configuration for the cluster.
+type KubernetesNetworkConfigResponse struct {
+	_ struct{} `type:"structure"`
+
+	// The CIDR block that Kubernetes service IP addresses are assigned from. If
+	// you didn't specify a CIDR block, then Kubernetes assigns addresses from either
+	// the 10.100.0.0/16 or 172.20.0.0/16 CIDR blocks. If this was specified, then
+	// it was specified when the cluster was created and it cannot be changed.
+	ServiceIpv4Cidr *string `locationName:"serviceIpv4Cidr" type:"string"`
+}
+
+// String returns the string representation
+func (s KubernetesNetworkConfigResponse) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s KubernetesNetworkConfigResponse) GoString() string {
+	return s.String()
+}
+
+// SetServiceIpv4Cidr sets the ServiceIpv4Cidr field's value.
+func (s *KubernetesNetworkConfigResponse) SetServiceIpv4Cidr(v string) *KubernetesNetworkConfigResponse {
+	s.ServiceIpv4Cidr = &v
+	return s
+}
+
 // An object representing a node group launch template specification. The launch
 // template cannot include SubnetId (https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_CreateNetworkInterface.html),
 // IamInstanceProfile (https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_IamInstanceProfile.html),
@@ -4997,7 +5081,7 @@ type Nodegroup struct {
 
 	// If the node group wasn't deployed with a launch template, then this is the
 	// disk size in the node group configuration. If the node group was deployed
-	// with a launch template, then diskSize is null.
+	// with a launch template, then this is null.
 	DiskSize *int64 `locationName:"diskSize" type:"integer"`
 
 	// The health status of the node group. If there are issues with your node group's
@@ -5006,7 +5090,7 @@ type Nodegroup struct {
 
 	// If the node group wasn't deployed with a launch template, then this is the
 	// instance type that is associated with the node group. If the node group was
-	// deployed with a launch template, then instanceTypes is null.
+	// deployed with a launch template, then this is null.
 	InstanceTypes []*string `locationName:"instanceTypes" type:"list"`
 
 	// The Kubernetes labels applied to the nodes in the node group.
@@ -5043,7 +5127,7 @@ type Nodegroup struct {
 
 	// If the node group wasn't deployed with a launch template, then this is the
 	// remote access configuration that is associated with the node group. If the
-	// node group was deployed with a launch template, then remoteAccess is null.
+	// node group was deployed with a launch template, then this is null.
 	RemoteAccess *RemoteAccessConfig `locationName:"remoteAccess" type:"structure"`
 
 	// The resources associated with the node group, such as Auto Scaling groups